applications/backend/lambda/authorizer/requirements.txt
# Authorizer Lambda 函数依赖
# 这是一个简单的授权函数，只需要最基本的依赖

# JWT验证
PyJWT==2.8.0

# AWS SDK (通常已内置，但显式声明)
boto3>=1.26.0